Fix miss match between generated test and results
It seems that the general agreement is that there should only be one system-out entry per test-suite:

http://help.catchsoftware.com/display/ET/JUnit+Format

So the original form is that there will be multiple system-out elements and for example Jenkins will only pick the first entry and ignore the rest. This fix means there is only one system-out element and all the attachments are shown.